+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add Mat3 into Rotor3 conversion for rotation matrices
 - Remove heavy reliance on `mul_add` due to negligible performance benefit and in many cases performance detriment.
 - Slightly optimize Vector `normalize`.
+- `Rotor2::from_angle_plane()` now takes plane and angle as separate arguments.
 
 ## 0.6.1
 - Add scalar multiplication and componentwise addition for `MatN`